fix: keep bucket-splitting in the workflow, out of a heavy module
The previous fix moved bucket-splitting to the client and changed the workflow to
take pre-split buckets, only to dodge a webpack bundling failure (the workflow had
a value import of period.ts, which pulls in Metronome and, transitively, the
Elasticsearch client). That changed the workflow's contract for a bundling reason,
not a logical one, and periods are still passed in unresolved by the caller as before.

Move splitConsumptionPeriodIntoBuckets to its own module with no heavy dependency
(just moment-timezone), so the workflow can import it as a value safely and go back
to taking `period` and chunking it internally.
