+++ front/temporal/analytics_queue/client.ts
@@ -1,5 +1,4 @@
 import type { ConsumptionPeriod } from "@app/lib/api/analytics/consumption/period";
-import { splitConsumptionPeriodIntoBuckets } from "@app/lib/api/analytics/consumption/period";
 import type { ConsumptionScopeFilter } from "@app/lib/api/analytics/consumption/scope";
 import type { Authenticator, AuthenticatorType } from "@app/lib/auth";
 import { getPrivateUploadBucket } from "@app/lib/file_storage";
@@ -268,13 +267,8 @@ export async function launchConsumptionExportWorkflow(
       return new Ok({ status: "already_running", workflowId, period, filter });
     }
 
-    // Computed here rather than inside the workflow: the workflow bundle must stay free
-    // of heavy runtime imports, and `period.ts` transitively pulls in the Elasticsearch
-    // client (through Metronome) which the Temporal workflow sandbox can't bundle.
-    const buckets = splitConsumptionPeriodIntoBuckets(exportPeriod);
-
     await client.workflow.start(runConsumptionExportWorkflow, {
-      args: [authType, { buckets, filter, exportId }],
+      args: [authType, { period: exportPeriod, filter, exportId }],
       taskQueue: QUEUE_NAME,
       workflowId,
       memo: {
